requirements.txt
streamlit
numpy
pillow
mtcnn
opencv-python-headless
scikit-learn
